api/requirements.txt
arrow
fastapi
gunicorn
loguru
notifiers
pydantic
pyotp
python-dotenv
python-multipart
uvicorn
web3
git+https://github.com/nftstorage/python-client.git